Fix generation of sand worm passable sand regions.
Also as result, this fix unfortunately uncovers some kind of bug in the rendering of the sand worm movement shimmer, which has some kind of blitting bug in it.

This causes the whole terrain layer to disappear whenever sand worm is moving, so disable the rendering of the shimmer for now.